Addressed some more warnings.

+++ S3.cpp
@@ -8,8 +8,8 @@ Mat4 cam_mat = Mat4::identity();
 
 Mat4 proj_mat = Mat4::identity();
 
-float aspect_ratio = 1;
-float fog_scale = 1.5;
+double aspect_ratio = 1;
+double fog_scale = 1.5;
 
 
 /*
