+++ S3.h
@@ -25,7 +25,7 @@
 extern Mat4 cam_mat;
 extern Mat4 proj_mat;
 
-extern float aspect_ratio, fog_scale;
+extern double aspect_ratio, fog_scale;
 
 void translate_cam(double right, double down, double fwd);
 void rotate_cam(double pitch, double yaw, double roll);
